requirements.txt
sqlalchemy~=2.0.44
python-dotenv~=1.2.1
Flask~=3.1.2
MarkupSafe~=3.0.3
Flask-Login~=0.6.3